+++ packages/plugins/src/cli/deploy.ts
@@ -30,7 +30,7 @@ import {
   writeDeploymentArtifacts,
   isLegacyTransactionsRequiredForNetwork,
   MAX_UINT64,
-  compileAndExecuteDeployment,
+  attemptDeployment,
   Deployment,
   fetchNameForNetwork,
   DeploymentContext,
@@ -427,7 +427,7 @@ export const deploy = async (
     provider,
     spinner,
   }
-  const result = await compileAndExecuteDeployment(deploymentContext)
+  const result = await attemptDeployment(deploymentContext)
 
   if (!result) {
     throw new Error(